Split send flow into 3 phases
